#17 Extract Instance.templateJson into TemplateRenderer class
