+++ server/src/test/scala/de/frosner/broccoli/models/InstanceSpec.scala
@@ -34,74 +34,6 @@ class InstanceSpec extends Specification with ScalaCheck with ModelArbitraries w
       Instance("1", Template("1", "\"{{id}}\"", "desc", Map.empty), Map.empty) must throwA[IllegalArgumentException]
     }
 
-    "parse the template correctly when it contains a single parameter" in {
-      val instance = Instance("1", Template("1", "\"{{id}}\"", "desc", Map.empty), Map("id" -> "Frank"))
-      instance.templateJson === JsString("Frank")
-    }
-
-    "parse the template correctly when it contains multiple parameters" in {
-      val instance =
-        Instance("1", Template("1", "\"{{id}} {{age}}\"", "desc", Map.empty), Map("id" -> "Frank", "age" -> "5"))
-      instance.templateJson === JsString("Frank 5")
-    }
-
-    "parse the template correctly when it contains a defined default parameter" in {
-      val instance = Instance(
-        id = "1",
-        template = Template(
-          id = "1",
-          template = "\"{{id}} {{age}}\"",
-          description = "desc",
-          parameterInfos = Map("age" -> ParameterInfo("age", None, Some("50"), secret = Some(false), `type` = None))
-        ),
-        parameterValues = Map("id" -> "Frank")
-      )
-      instance.templateJson === JsString("Frank 50")
-    }
-
-    "parse the template correctly when it has String parameters" in {
-      val instance = Instance(
-        id = "1",
-        template = Template(
-          id = "1",
-          template = "{{id}}",
-          description = "desc",
-          parameterInfos =
-            Map("id" -> ParameterInfo("id", None, None, secret = Some(false), `type` = Some(ParameterType.String)))
-        ),
-        parameterValues = Map("id" -> "Frank")
-      )
-      instance.templateJson === JsString("Frank")
-    }
-
-    "parse the template correctly when it contains regex stuff that breacks with replaceAll" in {
-      val instance = Instance(
-        id = "1",
-        template = Template(
-          id = "1",
-          template = "\"{{id}}\"",
-          description = "desc",
-          parameterInfos = Map.empty
-        ),
-        parameterValues = Map("id" -> "^.*$")
-      )
-      instance.templateJson === JsString("^.*$")
-    }
-
-    "parse the template correctly when it contains an undefined default parameter" in {
-      val instance = Instance(
-        id = "1",
-        template = Template(
-          id = "1",
-          template = "\"{{id}} {{age}}\"",
-          description = "desc",
-          parameterInfos = Map("age" -> ParameterInfo("age", None, None, secret = Some(false), `type` = None))
-        ),
-        parameterValues = Map("id" -> "Frank", "age" -> "50")
-      )
-      instance.templateJson === JsString("Frank 50")
-    }
-
     "throw an exception if the template contains no default and no value" in {
       Instance(
         id = "1",
